fix: cast substr position arg to integer in crockford_base32_encode
The remainder variable is BIGINT, so remainder + 1 produces BIGINT.
PostgreSQL's substr(text, integer, integer) rejects BIGINT for the
position argument, causing a type mismatch error when the trigger fires.

Cast (remainder + 1)::integer to match the expected signature.

+++ schema/tapdb_schema.sql
@@ -279,7 +279,7 @@ BEGIN
     END IF;
     WHILE val > 0 LOOP
         remainder := val % 32;
-        result := substr(alphabet, remainder + 1, 1) || result;
+        result := substr(alphabet, (remainder + 1)::integer, 1) || result;
         val := val / 32;
     END LOOP;
     RETURN result;
